+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
++ A new stack method based on bytecode prediction
+
 ## [0.1.2] - 2021-09-14
 
 ### Added