+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 + A new stack method based on bytecode prediction
 
+### Fixed
+
++ Generator states
+
 ## [0.1.2] - 2021-09-14
 
 ### Added