+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 + Generator states
 
+### Removed
+
++ CPython 3.8 support
+
 ## [0.1.2] - 2021-09-14
 
 ### Added